requirements.txt
colour
matplotlib
numpy
pandas
pillow
pytest
requests
scikit-image
scipy
torch
torchvision
tqdm
yt-dlp
pyyaml
rich
lightning
pydantic